Move batch struct conversion to models file
This moves the batch store struct conversion implementations to the
models.rs file. This is part of the batch store stabilization effort and
brings the store in line with the others.
